LPS-54221 move PortletRequestType to portal-service
