+++ portal-impl/src/com/liferay/portal/monitoring/statistics/portlet/PortletStatistics.java
@@ -17,6 +17,7 @@ package com.liferay.portal.monitoring.statistics.portlet;
 import com.liferay.portal.kernel.monitoring.MonitoringException;
 import com.liferay.portal.kernel.monitoring.RequestStatus;
 import com.liferay.portal.kernel.monitoring.statistics.DataSampleProcessor;
+import com.liferay.portal.kernel.monitoring.statistics.PortletRequestType;
 import com.liferay.portal.kernel.monitoring.statistics.RequestStatistics;
 
 import java.util.HashMap;
