+++ portal-service/src/com/liferay/portal/kernel/monitoring/statistics/PortletRequestType.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.monitoring.statistics.portlet;
+package com.liferay.portal.kernel.monitoring.statistics;
 
 /**
  * @author Michael C. Han
